+++ services/madoc-ts/src/frontend/site/features/canvas/CanvasTaskProgress.tsx
@@ -68,6 +68,11 @@ export const CanvasTaskProgress: React.FC = () => {
         status: 2,
         status_text: 'in progress',
       });
+      // mark parent as incomplete
+      await api.updateRevisionTask(canvasTask.parent_task, {
+        status: 2,
+        status_text: 'in progress',
+      });
       await refetch();
     }
   });